# Use string interning for fun and performance